One Page layout and Slideshow
Brought in most of the other pages and set them in the index for the one page layout, then added styles to give the newly added elements the spacing they needed. Redid the orange box shadow to a more dark shadow. Reworked the slideshow in CSS for desktop and I added arrows for the slideshow to use (not fully worked out)