Images, shop page, index file
Made many changes. Almost have the whole shopping page finished, it's responsive, having a 3 column grid layout for desktop and 1 column styles for mobile. Also changed the links for the shopping nav so that the only link is "Return Home". Javascript still needs to be added to it. Made some changes to the index file, specifically the CTSO. Finally, adjusted CSS styles, and removed some useless lines of code